Pass book progress and display it in BookPlay view
